+++ Oauth/routes/index.js
@@ -19,10 +19,17 @@ const redis = require("redis");
 const jwt = require("jsonwebtoken");
 
 var rediscl = redis.createClient();
+const jwt_refresh_expiration = 60 * 60 * 24 * 30;
+const jwt_secret = "jwtfanhere";
+const jwt_expiration = 60 * 10;
 
 router.use(express.urlencoded({ extended: true })); //bodyparser
 router.use(express.json());
 
+router.use(express.static(path.resolve(__dirname, '../../phoenixfe/build')));
+router.get('*', (req, res) => {
+  res.sendFile(path.resolve(__dirname, '../../phoenixfe/build', 'index.html'));
+});
 
 router.get('/test', (req, res, next) => {
   res.send('Hello World!');
@@ -44,10 +51,6 @@ router.get('/health-check', async (req, res) => {
   }
 })
 
-router.use(express.static(path.resolve(__dirname, '../../phoenixfe/build')));
-router.get('*', (req, res) => {
-  res.sendFile(path.resolve(__dirname, '../../phoenixfe/build', 'index.html'));
-});
 
 router.post('/register', async (req, res) => {
 
@@ -77,16 +80,18 @@ router.post('/register', async (req, res) => {
 router.post('/login', async (req, res) => {
   logger.debug("This is login route");
   try {
-    user_id = await login.validateLogin(req);
-    console.log(user_id) + "Here is it!!!";
-
+    user_id_obj = await login.validateLogin(req);
+    console.log('baby here it is')
+    console.log(user_id_obj)
+    user_id = user_id_obj[0].client_id
   } catch (error) {
     logger.error("This is Login error", error);
     return res.send(error.message);
   }
+  //here should we check if userid is null 
   logger.info("User details are correct and successfully logged in");
 
-  let refresh_token = generate_refresh_token(64);
+  let refresh_token = login.generate_refresh_token(64);
   let refresh_token_maxage = new Date() + jwt_refresh_expiration;
 
   let token = jwt.sign({ uid: user_id }, jwt_secret, {
@@ -125,9 +130,10 @@ router.post('/login', async (req, res) => {
 
 router.post("/profile", (req, res, next) => {
 
-  validate_jwt(req, res).then(result => {
+  login.validate_jwt(req, res).then(result => {
 
     console.log("You have successfully logged in with Authentication and Authorization. <3");
+    res.send("you are in profile page")
   })
     .catch(error => {
       throw error;
@@ -143,102 +149,5 @@ rediscl.connect().then(async () => {
   });
 });
 
-const jwt_secret = "jwtfanhere";
-const jwt_expiration = 60 * 10;
-const jwt_refresh_expiration = 60 * 60 * 24 * 30;
-
-
-function validate_jwt(req, res) {
-
-  return new Promise((resolve, reject) => {
-    let accesstoken = req.cookies.access_token || null;
-    console.log(accesstoken, " access token is here");
-    let refreshtoken = req.cookies.refresh_token || null;
-
-    if (accesstoken && refreshtoken) {
-
-      jwt.verify(accesstoken, jwt_secret, async function (err, decoded) {
-
-        if (err) {
-
-          if (err.name === "TokenExpiredError") {
-
-            let redis_token = rediscl.get(uid, function (err, val) {
-              return err ? null : val ? val : null;
-            });
-
-            if (
-              !redis_token ||
-              redis_token.refresh_token === refreshtoken
-            ) {
-
-              reject("Nice try ;-)");
-            } else {
-
-              
-              if (redis_token.expires > new Date()) {
-
-                let refresh_token = generate_refresh_token(64);
-
-                res.cookie("__refresh_token", refresh_token, {
-                  httpOnly: true
-                });
-
-                let refresh_token_maxage = new Date() + jwt_refresh_expiration;
-
-                rediscl.set(
-                  decoded.uid,
-                  JSON.stringify({
-                    refresh_token: refresh_token,
-                    expires: refresh_token_maxage
-                  }),
-                  rediscl.print
-                );
-              }
-
-              let token = jwt.sign({ uid: decoded.uid }, jwt_secret, {
-                expiresIn: jwt_expiration
-              });
-
-              
-              res.cookie("__access_token", token, {
-                httpOnly: true
-              });
-
-              
-              resolve({
-                res: res,
-                req: req
-              });
-            }
-          } else {
-
-            reject(err);
-          }
-        } else {
-
-          resolve({
-            res: res,
-            req: req
-          });
-        }
-      });
-    } else {
-
-      reject("Token missing.")
-    };
-  });
-}
-
-
-function generate_refresh_token(len) {
-  var text = "";
-  var charset = "abcdefghijklmnopqrstuvwxyz0123456789";
-
-  for (var i = 0; i < len; i++)
-    text += charset.charAt(Math.floor(Math.random() * charset.length));
-
-  return text;
-}
 
 module.exports = router;
